requirements.txt
discord.py==2.4.0
youtube-dl
PyNaCl
ffmpeg
python-dotenv
yt-dlp
imageio[ffmpeg]